+++ simulations-java/simulations/radio-waves/src/edu/colorado/phet/radiowaves/model/EmfSensingElectron.java
@@ -25,7 +25,7 @@ public class EmfSensingElectron extends PositionConstrainedElectron {
         super.setRecordHistory( false );
     }
 
-    private Vector2D.Float aPrev = new Vector2D.Float();
+    private Vector2D.Double aPrev = new Vector2D.Double();
 
     public synchronized void stepInTime( double dt ) {
         super.stepInTime( dt );
@@ -52,8 +52,8 @@ public class EmfSensingElectron extends PositionConstrainedElectron {
             else {
                 // The field strength is a force on the electron, so we must compute an
                 // acceleration
-                Vector2D.Float fieldStrength = sourceElectron.getDynamicFieldAt( location );
-                Vector2D.Float a = fieldStrength;
+                Vector2D.Double fieldStrength = sourceElectron.getDynamicFieldAt( location );
+                Vector2D.Double a = fieldStrength;
                 double x = this.getCurrentPosition().getX();
                 double y = this.getCurrentPosition().getY();
                 location = this.getCurrentPosition();
